+++ src/Gold.java
@@ -2,7 +2,7 @@
 public class Gold extends Customer{
     protected int discount;
 
-    Gold(String fn, String ln, String gi, double as, int d){
+    Gold(String fn, String ln, String gi, float as, int d){
         super(fn, ln, gi, as);
         discount = d;
     }
